==> code/3 human vs algorithm GAMs.Rmd <==
---
title: "human vs algorithm GAMs"
author: "LGCarlson"
date: "10/16/2020"
output: html_document
---

```{r setup, include=FALSE}
library(tidyverse)
library(lubridate)
library(janitor)
library(here)
library(mgcv)
library(tidymv)
library(patchwork)

check = function(x){summary(as.factor(x))}

#set personalized plotting theme
LGCtheme <- theme_light() + theme(panel.background = element_blank(), 
                panel.grid = element_blank(), 
                axis.text = element_text(size = 13), 
                axis.title = element_text(size = 14), 
                legend.text = element_text(size = 13), 
                legend.title = element_text(size = 14), 
                strip.background = element_blank(), 
                strip.text = element_text(size = 13, color = "black"))

theme_set(theme_light())
knitr::opts_chunk$set(echo = TRUE)
```

Read in human/algorithm data created in code 1

```{r,message=F}
alg_vs_human_markers<-read_csv(here::here("processed_data", "alg_vs_human_markers.csv")) %>% 
  mutate(markersource = as.factor(markersource)) 
```

Calculate AW.incr (after winter increment) for human marked

```{r}
human_afterwinter<-alg_vs_human_markers %>% 
  mutate(AW.incr = marine.incr - PS.incr) %>% 
  filter(markersource == "human") %>% 
  dplyr::select(JoinID,AW.incr) %>% 
  dplyr::rename("human" = "AW.incr")
```

Wrangle full dataset of algorithm placed markers, then join human data to use for detrending after winter increment
```{r}
detrend_df<-read_csv(here::here("raw_data", "algorithm_full_dataset.csv")) %>% 
  dplyr::rename("algorithm" = "afterwinter_incr") %>% 
  dplyr::select(JoinID, DaysatSea,algorithm) %>% 
  left_join(human_afterwinter,by="JoinID") %>% 
  gather(key="markersource", value ="AW.incr",3:4) %>% 
  filter(!is.na(AW.incr))
```

Use linear model to detrend after winter increment based on number of days at sea

```{r, fig.width=4,fig.height=6}
detrend_incr <- lm(detrend_df$AW.incr ~ detrend_df$DaysatSea)
detrend_df$DAW.incr <- resid(detrend_incr) + mean(detrend_df$AW.incr)

incr_unadjusted<-detrend_df %>% 
  ggplot(aes(DaysatSea, AW.incr)) + geom_jitter() + 
  geom_smooth(method = "lm", fill="#ABB400", color="#ABB400") + LGCtheme + 
  labs(y="Increment (mm)", x="Recapture julian date", title="Unadjusted postannulus growth")

incr_adjusted<-detrend_df %>% 
  ggplot(aes(DaysatSea, DAW.incr)) + geom_jitter() + 
  geom_smooth(method = "lm", fill="#ABB400", color="#ABB400") + LGCtheme + 
  labs(y="Increment (mm)", x="Recapture julian date", title="Adjusted postannulus growth")



adj_plots<-(incr_unadjusted /
  incr_adjusted)

adj_plots
```

Create a dataframe of for writing gams

```{r}
gam_df<-alg_vs_human_markers %>% 
  left_join(detrend_df,by=c("JoinID","markersource")) %>% 
  mutate(markersource = as.factor(markersource)) 

table(gam_df$ReleaseYear)
```

### First summer increment

```{r}
fs.gam2<-mgcv::gam(FS.incr ~ markersource+s(ReleaseYear,by=markersource, k=5)+s(smolt.incr, k=5),data=gam_df,select=T, method = 'REML')


fs_smooth<-plot_smooths(fs.gam2,ReleaseYear,markersource) +
  scale_color_manual(values=c("#00608A","#ABB400"),name = "Recapture", labels = c("At sea", "Homewater")) +
  scale_fill_manual(values=c("#00608A","#ABB400"),name = "Recapture", labels = c("At sea", "Homewater")) + 
  scale_linetype_discrete(name = "Recapture", labels = c("At sea", "Homewater")) + LGCtheme  +
  scale_y_continuous(limits = c(0.4,1.25)) + 
  theme(legend.position = "none") +  
  labs(x="Smolt Year", y = "First summer increment (mm)") +
  theme(axis.title.x.bottom = element_blank(), axis.text.x.bottom = element_blank(),
  axis.line.x.bottom = element_blank(), axis.ticks.x.bottom = element_blank())

fs_smooth_facet_rug<-plot_smooths(fs.gam2,ReleaseYear,markersource) +
  geom_rug(aes(x=ReleaseYear,y=FS.incr),data=alg_vs_human_markers, na.rm = T,alpha = 0.25, position = "jitter", sides="b",color="#ABB400") +
  #geom_rug(aes(x=ReleaseYear,y=FS.incr),color="#F4916B",data=hum_incrs, na.rm = T,alpha = 0.25, position = "jitter", sides="r") +
  #geom_rug(aes(x=ReleaseYear,y=FS.incr),color="#ABB400",data=alg_incrs, na.rm = T,alpha = 0.25, position = "jitter", sides="l") +
  scale_y_continuous(limits = c(0.4,1.25)) +
  scale_color_manual(values=c("#ABB400","#F4916B"),name = "", labels = c("Algorithm", "Human")) +
  scale_fill_manual(values=c("#ABB400","#F4916B"),name = "", labels = c("Algorithm", "Human")) + 
  scale_linetype_discrete(name = "", labels = c("Algorithm", "Human")) + LGCtheme  +
  theme(legend.position = "none") +  
  labs(x="", y="",title = "First summer increment (mm)") + 
  theme(panel.border = element_blank(),axis.line = element_line(color="grey"))


```

### First winter increment

```{r}

fw.gam2<-gam(FW.incr ~ markersource+s(ReleaseYear,by=markersource, k=5)+s(FS.incr, k=5),data=gam_df,select=T, method = 'REML')


fw_smooth<-plot_smooths(fw.gam2,ReleaseYear,markersource) +
  scale_color_manual(values=c("#00608A","#ABB400"),name = "Recapture", labels = c("At sea", "Homewater")) +
  scale_fill_manual(values=c("#00608A","#ABB400"),name = "Recapture", labels = c("At sea", "Homewater")) + 
  scale_linetype_discrete(name = "Recapture", labels = c("At sea", "Homewater")) + LGCtheme  +
  scale_y_continuous(limits = c(0.4,1.25)) + 
  theme(legend.position = "none") +  
  labs(x="Smolt Year", y = "First summer increment (mm)") +
  theme(axis.title.x.bottom = element_blank(), axis.text.x.bottom = element_blank(),
  axis.line.x.bottom = element_blank(), axis.ticks.x.bottom = element_blank())


fw_smooth_facet_rug<-plot_smooths(fw.gam2,ReleaseYear,markersource) +
  geom_rug(aes(x=ReleaseYear,y=FW.incr),data=alg_vs_human_markers, na.rm = T,alpha = 0.25, position = "jitter", sides="b",color="#ABB400") +
  #geom_rug(aes(x=ReleaseYear,y=FW.incr),color="#ABB400",data=hum_incrs, na.rm = T,alpha = 0.25, position = "jitter", sides="r") +
  #geom_rug(aes(x=ReleaseYear,y=FW.incr),color="#00608A",data=alg_incrs, na.rm = T,alpha = 0.25, position = "jitter", sides="l") +
  scale_y_continuous(limits = c(0.4,1.25)) +
  scale_color_manual(values=c("#ABB400","#F4916B"),name = "", labels = c("Algorithm", "Human")) +
  scale_fill_manual(values=c("#ABB400","#F4916B"),name = "", labels = c("Algorithm", "Human")) + 
  scale_linetype_discrete(name = "", labels = c("Algorithm", "Human")) + LGCtheme  +
  theme(legend.position = "bottom") +  
  labs(x="Smolt year", y="",title = "First winter increment (mm)") +
  theme(axis.title.y = element_blank(), axis.line = element_line(color="grey"),
        panel.border = element_blank())

```


### Second year increment 

```{r}

y2.gam2<-gam(DAW.incr ~ markersource+s(ReleaseYear,by=markersource, k=2)+s(PS.incr, k=5),data=gam_df,select=T, method = 'REML')

# *uses detrended after winter increment rather

y2_smooth<-plot_smooths(y2.gam2,ReleaseYear,markersource) +
  scale_y_continuous(limits = c(0.4,1.25)) +
  LGCtheme  +
  theme(legend.position = "bottom") +  
  labs(x="Smolt Year", y = "Second year increment (mm)")



y2_smooth_facet_rug<-plot_smooths(y2.gam2,ReleaseYear,markersource) +
  geom_rug(aes(x=ReleaseYear,y=Y2.incr),data=alg_vs_human_markers, na.rm = T,alpha = 0.25, position = "jitter", sides="b",color="#ABB400") +
  #geom_rug(aes(x=ReleaseYear,y=Y2.incr),color="#ABB400",data=hum_incrs, na.rm = T,alpha = 0.25, position = "jitter", sides="r") +
  #geom_rug(aes(x=ReleaseYear,y=Y2.incr),color="#00608A",data=alg_incrs, na.rm = T,alpha = 0.25, position = "jitter", sides="l") +
  scale_y_continuous(limits = c(0.4,1.25)) +
  scale_color_manual(values=c("#ABB400","#F4916B"),name = "", labels = c("Algorithm", "Human")) +
  scale_fill_manual(values=c("#ABB400","#F4916B"),name = "", labels = c("Algorithm", "Human")) + 
  scale_linetype_discrete(name = "", labels = c("Algorithm", "Human")) + LGCtheme  +
  theme(legend.position = "none") +  
  labs(x="", y = "", title = "Second year increment (mm)")  +
  theme(axis.title.y = element_blank(), axis.line = element_line(color="grey"),
        panel.border = element_blank())
y2_smooth_facet_rug
```


```{r, fig.height=4.25, fig.width=10}
seasonal_GAMs<-fs_smooth_facet_rug +   fw_smooth_facet_rug +   y2_smooth_facet_rug

seasonal_GAMs 
```



```{r, warning=F, message=F, fig.width=10, fig.height=4.25}

freshxfs<-plot_smooths(fs.gam2,smolt.incr,markersource) + 
  scale_x_continuous(limits=c(0.5,2)) + scale_y_continuous(limits = c(0,1.5)) +
  scale_color_manual(values=c("#ABB400","#F4916B")) +
  scale_fill_manual(values=c("#ABB400","#F4916B")) + 
  LGCtheme  + theme(legend.position = "none") +  
  labs(x="Freshwater inc. (mm)", y = "First summer inc. (mm)") +
  theme(panel.border = element_blank(),axis.line = element_line(color="grey"))


fsxfw<-plot_smooths(fw.gam2,FS.incr,markersource) +
  scale_x_continuous(limits=c(0,2)) + scale_y_continuous(limits = c(0,1.5)) +
  scale_color_manual(values=c("#ABB400","#F4916B"),name = "", labels = c("Algorithm", "Human")) +
  scale_fill_manual(values=c("#ABB400","#F4916B"),name = "", labels = c("Algorithm", "Human")) + 
  scale_linetype_discrete(name = "", labels = c("Algorithm", "Human")) +
  LGCtheme  + theme(legend.position = "bottom") +  
  labs(x="First summer inc. (mm)", y = "First winter inc. (mm)") +
  theme(panel.border = element_blank(),axis.line = element_line(color="grey")) 



psxy2<-plot_smooths(y2.gam2,PS.incr,markersource) + 
  scale_x_continuous(limits=c(0.75,1.75)) + scale_y_continuous(limits = c(0,1.5)) +
  scale_color_manual(values=c("#ABB400","#F4916B")) +
  scale_fill_manual(values=c("#ABB400","#F4916B")) + 
  LGCtheme  + theme(legend.position = "none") +  
  labs(x="Postsmolt inc. (mm)", y = "Second year inc. (mm)") +
  theme(panel.border = element_blank(),axis.line = element_line(color="grey"))

condeffect<-(freshxfs + fsxfw + psxy2)
condeffect
```



```{r, fig.height=7.5, fig.width=10}
seasonal_GAMs2<-(fs_smooth_facet_rug + xlab(label = "Smolt year") + fw_smooth_facet_rug + theme(legend.position = "none") + y2_smooth_facet_rug + xlab(label = "Smolt year"))/
  (freshxfs + fsxfw + psxy2)

seasonal_GAMs2

```
